# Improve update_ancestors_after_delete trigger