Previous benchmark results:
```
delete_node_worst_case        4.00s
delete_node_best_case         1.32s
```

New benchmark results:
```
delete_node_worst_case        1.98s
delete_node_best_case         317.17ms
```

Postgresql 9.4.5 with default settings on OSX.