@vortec 2015-10-27
Thank you, that's a nice speed-up! I get similar results, before:

```
delete_node_worst_case        5.16s
delete_node_best_case         1.45s
```

After:

```
delete_node_worst_case        3.97s
delete_node_best_case         327.91ms
```